mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.2",
    "beautifulsoup4>=4.13.3",
    "python-dotenv>=1.0.1",
    "openai>=1.55.1",
]
